refactor(dashboard): share frontmatter validation
